Manually generate requirements.txt
The pipenv-to-requirements package was generating two separate files for
Pipenv packages and dev-packages, and only one of these could be passed
to Read the Docs. This was causing the build on Read the Docs to fail
because some packages were not installed.

Manually generating requirements.txt all the dev packages should fix
this issues.